+++ fut/core.py
@@ -215,10 +215,10 @@ class Core(object):
         # emulate
         if emulate == 'ios':
             sku = 'FUT16IOS'
-            clientVersion = 17
+            clientVersion = 20
         elif emulate == 'and':
             sku = 'FUT16AND'
-            clientVersion = 17
+            clientVersion = 20
 #        TODO: need more info about log in procedure in game
 #        elif emulate == 'xbox':
 #            sku = 'FFA16XBX'  # FFA14CAP ?